+++ src/util.cs
@@ -631,7 +631,7 @@ static public class AST_Util
     n.name = name;
     n.scope_type = scope_type;
     n.line_num = line_num;
-    n.symb_idx = (uint)symb_idx;
+    n.symb_idx = symb_idx;
     n.module_id = module_id;
 
     return n;
